Core/DataStores: Updated db2 structures to 9.2.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/7fbdd6473e5c67d002029d3c3675977e6b037e88)

+++ Source/Framework/Constants/CliDBConst.cs
@@ -1798,7 +1798,7 @@ namespace Framework.Constants
         DontDismissWhenEncounterIsAborted = 0x80000000  // NYI
     }
 
-    public enum TaxiNodeFlags : byte
+    public enum TaxiNodeFlags : ushort
     {
         Alliance = 0x1,
         Horde = 0x2,
